+++ cmd/diagnostics.go
@@ -21,7 +21,7 @@ var diagnosticsCmd = &cobra.Command{
 		if common.DebugFlag {
 			defer common.Stopwatch("Diagnostic run lasted").Report()
 		}
-		_, err := operations.ProduceDiagnostics(fileOption, robotOption, jsonFlag)
+		_, err := operations.ProduceDiagnostics(fileOption, robotOption, jsonFlag, productionFlag)
 		if err != nil {
 			pretty.Exit(1, "Error: %v", err)
 		}
@@ -34,4 +34,5 @@ func init() {
 	diagnosticsCmd.Flags().BoolVarP(&jsonFlag, "json", "j", false, "Output in JSON format")
 	diagnosticsCmd.Flags().StringVarP(&fileOption, "file", "f", "", "Save output into a file.")
 	diagnosticsCmd.Flags().StringVarP(&robotOption, "robot", "r", "", "Full path to 'robot.yaml' configuration file. [optional]")
+	diagnosticsCmd.Flags().BoolVarP(&productionFlag, "production", "p", false, "Checks for production level robots. [optional]")
 }