+++ operations/issues.go
@@ -60,7 +60,7 @@ func createIssueZip(attachmentsFiles []string) (string, error) {
 
 func createDiagnosticsReport(robotfile string) (string, *common.DiagnosticStatus, error) {
 	file := filepath.Join(conda.RobocorpTemp(), "diagnostics.txt")
-	diagnostics, err := ProduceDiagnostics(file, robotfile, false)
+	diagnostics, err := ProduceDiagnostics(file, robotfile, false, false)
 	if err != nil {
 		return "", nil, err
 	}